Separates the savings goal functions from functions.rb and moves them to their own file.
